Moved to a  profile util folder
