+++ src/profile_utils/userLicenseUtils.ts
